requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.1
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
six==1.16.0
sympy==1.12
torch==1.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
